backend/requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
docx2pdf
python-pptx